requirements.txt
requests
Flask
openslide-python
numpy
pyautogui
psutil